@@ -17,42 +17,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-mdc/web": "^0.44.0",
+    "@angular-mdc/web": "^2.0.1",
     "@angular-persian/material-date-picker": "^1.0.3",
-    "@angular/animations": "^7.2.7",
-    "@angular/bazel": "7.2.15",
-    "@angular/cdk": "^7.3.3",
-    "@angular/common": "^7.2.7",
-    "@angular/compiler": "^7.2.7",
-    "@angular/core": "^7.2.7",
-    "@angular/forms": "^7.2.7",
-    "@angular/http": "^7.2.7",
-    "@angular/material": "^7.3.3",
-    "@angular/material-moment-adapter": "^7.3.3",
-    "@angular/platform-browser": "^7.2.7",
-    "@angular/platform-browser-dynamic": "^7.2.7",
-    "@angular/platform-server": "^7.2.7",
-    "@angular/pwa": "0.12",
-    "@angular/router": "^7.2.7",
-    "@angular/service-worker": "^7.2.7",
-    "@material/tab-bar": "^1.1.0",
+    "@angular/animations": "^8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "^8.1.0",
+    "@angular/compiler": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/forms": "^8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/material-moment-adapter": "^8.0.2",
+    "@angular/platform-browser": "^8.1.0",
+    "@angular/platform-browser-dynamic": "^8.1.0",
+    "@angular/platform-server": "^8.1.0",
+    "@angular/pwa": "0.801.0",
+    "@angular/router": "^8.1.0",
+    "@angular/service-worker": "^8.1.0",
+    "@material/tab-bar": "^3.0.0",
     "@ng-stack/contenteditable": "^1.0.1",
-    "@ng-toolkit/universal": "^1.1.47",
-    "@nguniversal/express-engine": "^7.0.2",
-    "@nguniversal/module-map-ngfactory-loader": "^7.0.2",
-    "@swimlane/ngx-charts": "^10.1.0",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "^8.1.1",
     "@types/chart.js": "^2.7.46",
-    "@types/lodash": "^4.14.118",
+    "@types/lodash": "^4.14.136",
     "@types/notyf": "^2.0.0",
     "@types/socket.io-client": "^1.4.32",
     "angular2-chartjs": "^0.5.1",
     "angular2-persian-pipes": "^0.0.6",
-    "apexcharts": "^3.6.2",
+    "apexcharts": "^3.8.2",
     "audio-loader": "^1.0.3",
     "audio-play": "^2.2.1",
     "audit": "0.0.6",
-    "chartist": "^0.11.0",
-    "core-js": "^2.5.7",
+    "chartist": "^0.11.3",
+    "core-js": "^3.1.4",
     "cors": "~2.8.4",
     "dexie": "^2.0.4",
     "file-saver": "^2.0.0",
@@ -79,31 +75,30 @@
     "push-js": "^1.1.1",
     "push.js": "^1.0.9",
     "reconnecting-websocket": "^4.1.10",
-    "rxjs": "6.4.0",
+    "rxjs": "6.5.2",
     "rxjs-compat": "^6.3.2",
     "socket.io-client": "^2.1.1",
     "soso-client": "^3.2.0",
-    "ts-loader": "5.3.3",
-    "typescript": "3.2.4",
+    "ts-loader": "6.0.4",
     "webpack-cli": "^3.2.3",
     "xlsx": "^0.14.1",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "~0.13.4",
-    "@angular/cli": "^7.3.4",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "^7.2.7",
-    "@bazel/ibazel": "^0.9.1",
+    "@angular-devkit/build-angular": "~0.801.0",
+    "@angular-devkit/build-ng-packagr": "~0.801.0",
+    "@angular/cli": "^8.1.0",
+    "@angular/compiler-cli": "^8.1.0",
+    "@angular/language-service": "^8.1.0",
+    "@bazel/ibazel": "^0.10.3",
     "@types/chartist": "^0.9.44",
     "@types/jasmine": "~3.3.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^11.10.5",
-    "codelyzer": "~4.5.0",
-    "electron": "^4.1.4",
+    "@types/node": "^12.6.1",
+    "codelyzer": "^5.0.1",
+    "electron": "^5.0.6",
     "electron-builder": "^20.41.0",
-    "jasmine-core": "~3.3.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.0.1",
     "karma-chrome-launcher": "~2.2.0",
@@ -113,11 +108,11 @@
     "ng-packagr": "^5.1.0",
     "persianjs": "^0.4.0",
     "protractor": "^5.4.1",
-    "ts-node": "~8.0.2",
-    "tsickle": "^0.35.0",
-    "tslib": "^1.9.3",
-    "tslint": "~5.13.1",
-    "webpack": "^4.31.0"
+    "ts-node": "~8.3.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.18.0",
+    "typescript": "3.4.5",
+    "webpack": "^4.35.3"
   },
   "build": {
     "appId": "raydaDesktop",
