@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers } from '@angular/http';
+import { HttpClient, HttpHeaders  } from '@angular/common/http';
 import { BehaviorSubject } from 'rxjs';
 import { environment } from '../../environments/environment';
 
@@ -7,7 +7,7 @@ import { environment } from '../../environments/environment';
   providedIn: 'root'
 })
 export class AuthService {
-  constructor(private http: Http) {
+  constructor(private http: HttpClient) {
 
   }
   private messageSource = new BehaviorSubject('hello');
@@ -19,20 +19,20 @@ export class AuthService {
   login(data) {
     return this.http.post(environment.UserApi + 'login', data).toPromise()
     .then((d) => {
-       if (d.json().status) {
+       if (d['status']) {
          this.changeMessage('logged');
-       localStorage.setItem('token', d.json().token);
+       localStorage.setItem('token', d['token']);
        localStorage.setItem('uid', data.username);
 
        }
-       return d.json().status;
+       return d['status'];
 
     });
   }
   newpass(data) {
     return this.http.post(environment.UserApi + 'new', data).toPromise()
     .then((d) => {
-       return d.json().status;
+       return d['status'];
 
     });
   }
@@ -41,17 +41,19 @@ export class AuthService {
     .toPromise();
   }
   async checkLevel() {
-    const headers = new Headers();
-    headers.append('Authorization', 'Bearer ' + localStorage.token);
-    try {
-      const ok = await this.http.get('https://levelc.rayda.ir/' + `level?q=${Math.random()}`, {
-        headers : headers
+    const httpOptions = {
+      headers: new HttpHeaders({
+        'Content-Type':  'application/json',
+        'Authorization': 'Bearer ' + localStorage.token
       })
+    };
+    try {
+      const ok = await this.http.get('https://levelc.rayda.ir/' + `level?q=${Math.random()}`, httpOptions)
       .toPromise();
 
 
 
-      return ok.json().level;
+      return ok['level'];
     } catch (error) {
       console.log('check level :'  + error);
     }
