@@ -1,4 +1,4 @@
-import { DateAdapter } from "@angular/material";
+import { DateAdapter } from "@angular/material/core";
 import * as jalaliMoment from "jalali-moment";
 
 export const PERSIAN_DATE_FORMATS = {
